refactor(server): decompose routes/routines.rs into child modules (#4039)
Split the 1032-prod-line route file along its natural seams into private
child modules — handlers.rs (route handlers), helpers.rs (validation/
runtime-runnable checks), audit.rs (audit logging), responses.rs (response
mapping) — keeping the facade's public handler API and router wiring
byte-identical. Pure move + pub(super) visibility; no logic change.

Also un-ignores src/server/routes/routines/ in .gitignore: the repo-wide
'routines/' per-user-script ignore would otherwise swallow the new module
directory.

Removes the giant_file_registry.toml entry per registry rules (all
resulting files < 1000 prod lines: 154/495/244/115/76).

Implemented-by: Codex
