+++ src/server/routes/routines.rs
@@ -1,25 +1,8 @@
-use axum::{
-    Json,
-    extract::{Path, Query, State},
-    http::{HeaderMap, StatusCode},
-};
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Deserializer};
-use serde_json::{Value, json};
+use serde_json::Value;
 
-use crate::config::Config;
-use crate::error::{AppError, AppResult, ErrorCode};
-use crate::services::routines::{
-    DeleteRoutineResult, NewRoutine, RoutineAgentExecutor, RoutineDiscordLogger,
-    RoutineLifecycleEvent, RoutinePatch, RoutineScriptLoader, RoutineSessionCommand,
-    RoutineSessionController, RoutineStore, execute_claimed_script_run,
-    is_migrated_launchd_script_ref, is_resume_routine_requires_next_due_at,
-    validate_migrated_launchd_activation, validate_routine_runtime_config,
-    validate_routine_schedule,
-};
-use crate::utils::api::clamp_api_limit;
-
-use super::AppState;
+use crate::services::routines::RoutinePatch;
 
 const PARALLEL_SAFE_MIGRATED_LAUNCHD_SCRIPT_REF: &str = "migrated-launchd/queue-stability-batch.js";
 
@@ -154,937 +137,26 @@ impl PatchRoutineBody {
     }
 }
 
-pub async fn list_routines(
-    State(state): State<AppState>,
-    Query(query): Query<ListRoutinesQuery>,
-) -> AppResult<Json<Value>> {
-    let store = routine_store(&state)?;
-    let routines = store
-        .list_routines(query.agent_id.as_deref(), query.status.as_deref())
-        .await
-        .map_err(store_error)?;
-    Ok(Json(json!({
-        "routines": routines,
-        "default_timeout_secs": state.config.routines.agent_timeout_secs,
-    })))
-}
+#[path = "routines/audit.rs"]
+mod audit;
+#[path = "routines/handlers.rs"]
+mod handlers;
+#[path = "routines/helpers.rs"]
+mod helpers;
+#[path = "routines/responses.rs"]
+mod responses;
 
-pub async fn routine_metrics(
-    State(state): State<AppState>,
-    Query(query): Query<RoutineMetricsQuery>,
-) -> AppResult<Json<Value>> {
-    let store = routine_store(&state)?;
-    let metrics = store
-        .metrics(query.agent_id.as_deref(), query.since)
-        .await
-        .map_err(store_error)?;
-    Ok(Json(json!({
-        "metrics": metrics,
-        "filters": {
-            "agent_id": query.agent_id,
-            "since": query.since,
-        },
-    })))
-}
-
-pub async fn search_routine_run_results(
-    State(state): State<AppState>,
-    Query(query): Query<SearchRoutineRunResultsQuery>,
-) -> AppResult<Json<Value>> {
-    let q = query.q.trim();
-    if q.is_empty() {
-        return Err(AppError::bad_request("q is required"));
-    }
-    if let Some(status) = query.status.as_deref() {
-        validate_run_status_filter(status)?;
-    }
-    let store = routine_store(&state)?;
-    let limit = clamp_api_limit(Some(query.limit.unwrap_or(20).max(0) as usize)) as i64;
-    let runs = store
-        .search_run_results(
-            q,
-            query.agent_id.as_deref(),
-            query.status.as_deref(),
-            query.since,
-            limit,
-        )
-        .await
-        .map_err(store_error)?;
-    Ok(Json(json!({
-        "runs": runs,
-        "filters": {
-            "q": q,
-            "agent_id": query.agent_id,
-            "status": query.status,
-            "since": query.since,
-            "limit": limit,
-        },
-    })))
-}
-
-pub async fn get_routine(
-    State(state): State<AppState>,
-    Path(routine_id): Path<String>,
-) -> AppResult<Json<Value>> {
-    let store = routine_store(&state)?;
-    let Some(routine) = store.get_routine(&routine_id).await.map_err(store_error)? else {
-        return Err(AppError::not_found(format!(
-            "routine {routine_id} not found"
-        )));
-    };
-    Ok(Json(json!({
-        "routine": routine,
-        "default_timeout_secs": state.config.routines.agent_timeout_secs,
-    })))
-}
-
-pub async fn list_routine_runs(
-    State(state): State<AppState>,
-    Path(routine_id): Path<String>,
-    Query(query): Query<ListRunsQuery>,
-) -> AppResult<Json<Value>> {
-    let store = routine_store(&state)?;
-    if store
-        .get_routine(&routine_id)
-        .await
-        .map_err(store_error)?
-        .is_none()
-    {
-        return Err(AppError::not_found(format!(
-            "routine {routine_id} not found"
-        )));
-    }
-    let runs = store
-        .list_runs(&routine_id, query.limit.unwrap_or(20))
-        .await
-        .map_err(store_error)?;
-    Ok(Json(json!({ "runs": runs })))
-}
-
-pub async fn attach_routine(
-    State(state): State<AppState>,
-    Json(body): Json<AttachRoutineBody>,
-) -> AppResult<(StatusCode, Json<Value>)> {
-    let store = routine_store(&state)?;
-    let script_ref = normalize_script_ref(&body.script_ref)?;
-    let name = body.name.unwrap_or_else(|| fallback_name(&script_ref));
-    let execution_strategy = body
-        .execution_strategy
-        .unwrap_or_else(|| "fresh".to_string());
-    validate_execution_strategy_request(&execution_strategy)?;
-    validate_schedule_request(body.schedule.as_deref())?;
-    validate_timeout_request(body.timeout_secs)?;
-    validate_max_retries_request(body.max_retries)?;
-    let agent_id = validate_agent_id_request(&state, "agent_id", body.agent_id.as_deref()).await?;
-    let fallback_agent_id = validate_agent_id_request(
-        &state,
-        "fallback_agent_id",
-        body.fallback_agent_id.as_deref(),
-    )
-    .await?;
-    validate_distinct_fallback_agent(agent_id.as_deref(), fallback_agent_id.as_deref())?;
-    let initial_status = initial_attach_status(&script_ref).to_string();
-    let routine = store
-        .attach_routine(NewRoutine {
-            agent_id,
-            fallback_agent_id,
-            max_retries: body.max_retries,
-            script_ref,
-            name,
-            status: Some(initial_status),
-            execution_strategy,
-            schedule: body.schedule,
-            next_due_at: body.next_due_at,
-            checkpoint: body.checkpoint,
-            discord_thread_id: body.discord_thread_id,
-            timeout_secs: body.timeout_secs,
-        })
-        .await
-        .map_err(store_error)?;
-    let discord_log = routine_discord_logger(&state)?
-        .log_routine_event_with_store(&store, &routine, RoutineLifecycleEvent::Attached)
-        .await;
-    Ok((
-        StatusCode::CREATED,
-        Json(json!({ "routine": routine, "discord_log": discord_log })),
-    ))
-}
-
-pub async fn patch_routine(
-    State(state): State<AppState>,
-    Path(routine_id): Path<String>,
-    Json(body): Json<PatchRoutineBody>,
-) -> AppResult<Json<Value>> {
-    let store = routine_store(&state)?;
-    if let Some(strategy) = body.execution_strategy.as_deref() {
-        validate_execution_strategy_request(strategy)?;
-    }
-    if let Some(Some(schedule)) = body.schedule.as_present() {
-        validate_schedule_request(Some(schedule))?;
-    }
-    if let Some(timeout_secs) = body.timeout_secs.as_present().copied().flatten() {
-        validate_timeout_request(Some(timeout_secs))?;
-    }
-    validate_max_retries_request(body.max_retries)?;
-    if let Some(fallback_agent_id) = body.fallback_agent_id.as_present() {
-        let fallback_agent_id =
-            validate_agent_id_request(&state, "fallback_agent_id", fallback_agent_id.as_deref())
-                .await?;
-        let current = store
-            .get_routine(&routine_id)
-            .await
-            .map_err(store_error)?
-            .ok_or_else(|| AppError::not_found(format!("routine {routine_id} not found")))?;
-        validate_distinct_fallback_agent(
-            current.agent_id.as_deref(),
-            fallback_agent_id.as_deref(),
-        )?;
-    }
-    let patch = body.into_patch();
-    let Some(routine) = store
-        .patch_routine(&routine_id, patch)
-        .await
-        .map_err(store_error)?
-    else {
-        return Err(AppError::not_found(format!(
-            "routine {routine_id} not found"
-        )));
-    };
-    Ok(Json(json!({ "routine": routine })))
-}
-
-pub async fn pause_routine(
-    State(state): State<AppState>,
-    Path(routine_id): Path<String>,
-) -> AppResult<Json<Value>> {
-    let store = routine_store(&state)?;
-    let changed = store
-        .pause_routine(&routine_id)
-        .await
-        .map_err(store_error)?;
-    if !changed {
-        return Err(AppError::not_found(format!(
-            "enabled routine {routine_id} not found"
-        )));
-    }
-    let discord_log = routine_discord_logger(&state)?
-        .log_routine_event_by_id(&store, &routine_id, RoutineLifecycleEvent::Paused)
-        .await;
-    Ok(Json(
-        json!({ "ok": true, "routine_id": routine_id, "discord_log": discord_log }),
-    ))
-}
-
-pub async fn resume_routine(
-    State(state): State<AppState>,
-    Path(routine_id): Path<String>,
-    Json(body): Json<ResumeRoutineBody>,
-) -> AppResult<Json<Value>> {
-    let store = routine_store(&state)?;
-    let Some(routine) = store.get_routine(&routine_id).await.map_err(store_error)? else {
-        return Err(AppError::not_found(format!(
-            "paused routine {routine_id} not found"
-        )));
-    };
-    if routine.status != "paused" {
-        return Err(AppError::not_found(format!(
-            "paused routine {routine_id} not found"
-        )));
-    }
-    let metadata = migrated_launchd_metadata_for_state(&state, &routine.script_ref)?;
-    let routine_script_dirs = state.config.routines.script_dirs();
-    validate_migrated_launchd_activation(
-        &routine.script_ref,
-        routine.checkpoint.as_ref(),
-        metadata.as_ref(),
-        &routine_script_dirs,
-    )?;
-    let changed = store
-        .resume_routine(&routine_id, body.next_due_at_update())
-        .await
-        .map_err(store_error)?;
-    if !changed {
-        return Err(AppError::not_found(format!(
-            "paused routine {routine_id} not found"
-        )));
-    }
-    let discord_log = routine_discord_logger(&state)?
-        .log_routine_event_by_id(&store, &routine_id, RoutineLifecycleEvent::Resumed)
-        .await;
-    Ok(Json(
-        json!({ "ok": true, "routine_id": routine_id, "discord_log": discord_log }),
-    ))
-}
-
-pub async fn detach_routine(
-    State(state): State<AppState>,
-    Path(routine_id): Path<String>,
-) -> AppResult<Json<Value>> {
-    let store = routine_store(&state)?;
-    let changed = store
-        .detach_routine(&routine_id)
-        .await
-        .map_err(store_error)?;
-    if !changed {
-        return Err(AppError::conflict(format!(
-            "routine {routine_id} is missing, already detached, or currently running"
-        )));
-    }
-    let discord_log = routine_discord_logger(&state)?
-        .log_routine_event_by_id(&store, &routine_id, RoutineLifecycleEvent::Detached)
-        .await;
-    Ok(Json(
-        json!({ "ok": true, "routine_id": routine_id, "discord_log": discord_log }),
-    ))
-}
-
-pub async fn delete_routine(
-    State(state): State<AppState>,
-    Path(routine_id): Path<String>,
-    headers: HeaderMap,
-) -> AppResult<(StatusCode, Json<Value>)> {
-    let store = routine_store(&state)?;
-    let caller_agent_id =
-        crate::services::kanban::resolve_requesting_agent_id_with_pg(store.pool(), &headers).await;
-    let result = store
-        .delete_detached_routine(&routine_id, caller_agent_id.as_deref())
-        .await
-        .map_err(store_error)?;
-    audit_routine_delete(&routine_id, &result);
-    delete_routine_response(&routine_id, result)
-}
-
-pub async fn run_routine_now(
-    State(state): State<AppState>,
-    Path(routine_id): Path<String>,
-) -> AppResult<Json<Value>> {
-    ensure_routine_runtime_runnable(&state.config.routines)?;
-
-    let store = routine_store(&state)?;
-    let Some(routine) = store.get_routine(&routine_id).await.map_err(store_error)? else {
-        return Err(AppError::not_found(format!(
-            "routine {routine_id} not found"
-        )));
-    };
-
-    let loader = RoutineScriptLoader::new().map_err(|error| {
-        AppError::internal(format!("routine script loader init failed: {error}"))
-            .with_code(ErrorCode::Internal)
-    })?;
-    let routine_script_dirs = state.config.routines.script_dirs();
-    loader.load_dirs(&routine_script_dirs).map_err(|error| {
-        AppError::internal(format!("routine script registry load failed: {error}"))
-            .with_code(ErrorCode::Config)
-    })?;
-    let metadata = if is_migrated_launchd_script_ref(&routine.script_ref) {
-        let Some(script) = loader.get_script(&routine.script_ref).map_err(|error| {
-            AppError::internal(format!("routine script lookup failed: {error}"))
-                .with_code(ErrorCode::Config)
-        })?
-        else {
-            return Err(AppError::conflict(format!(
-                "migrated routine {} is invalid: routine script not loaded",
-                routine.script_ref
-            )));
-        };
-        Some(script.metadata)
-    } else {
-        None
-    };
-    validate_migrated_launchd_activation(
-        &routine.script_ref,
-        routine.checkpoint.as_ref(),
-        metadata.as_ref(),
-        &routine_script_dirs,
-    )?;
-
-    let Some(claimed) = store
-        .claim_run_now(&routine_id)
-        .await
-        .map_err(store_error)?
-    else {
-        return Err(AppError::conflict(format!(
-            "routine {routine_id} is not enabled or already running"
-        )));
-    };
-
-    let agent_executor = routine_agent_executor(&state)?;
-    let discord_logger = routine_discord_logger(&state)?;
-    discord_logger.log_run_started(&store, &claimed).await;
-    let run_id = claimed.run_id.clone();
-    let pause_on_terminal_failure = state.config.routines.failure_pause_auto_resume_secs > 0;
-    let Some(outcome) = execute_claimed_script_run(
-        &store,
-        &loader,
-        Some(&agent_executor),
-        Some(&discord_logger),
-        claimed,
-        pause_on_terminal_failure,
-    )
-    .await
-    .map_err(store_error)?
-    else {
-        return Err(AppError::conflict(format!(
-            "routine run {run_id} was already closed before outcome capture"
-        )));
-    };
-    let discord_log = discord_logger.log_run_outcome(&store, &outcome).await;
-    Ok(Json(
-        json!({ "outcome": outcome, "discord_log": discord_log }),
-    ))
-}
-
-fn ensure_routine_runtime_runnable(config: &crate::config::RoutinesConfig) -> AppResult<()> {
-    if !config.enabled {
-        return Err(AppError::new(
-            StatusCode::SERVICE_UNAVAILABLE,
-            ErrorCode::Config,
-            "routines are disabled by config",
-        ));
-    }
-    if let Err(error) = validate_routine_runtime_config(config) {
-        return Err(AppError::new(
-            StatusCode::SERVICE_UNAVAILABLE,
-            ErrorCode::Config,
-            format!("routine runtime is not runnable: {}", error.message()),
-        ));
-    }
-    Ok(())
-}
-
-fn migrated_launchd_metadata_for_state(
-    state: &AppState,
-    script_ref: &str,
-) -> AppResult<Option<Value>> {
-    if !is_migrated_launchd_script_ref(script_ref) {
-        return Ok(None);
-    }
-    let loader = RoutineScriptLoader::new().map_err(|error| {
-        AppError::internal(format!("routine script loader init failed: {error}"))
-            .with_code(ErrorCode::Internal)
-    })?;
-    let routine_script_dirs = state.config.routines.script_dirs();
-    loader.load_dirs(&routine_script_dirs).map_err(|error| {
-        AppError::internal(format!("routine script registry load failed: {error}"))
-            .with_code(ErrorCode::Config)
-    })?;
-    let Some(script) = loader.get_script(script_ref).map_err(|error| {
-        AppError::internal(format!("routine script lookup failed: {error}"))
-            .with_code(ErrorCode::Config)
-    })?
-    else {
-        return Err(AppError::conflict(format!(
-            "migrated routine {script_ref} is invalid: routine script not loaded"
-        )));
-    };
-    Ok(Some(script.metadata))
-}
-
-pub async fn reset_routine_session(
-    State(state): State<AppState>,
-    Path(routine_id): Path<String>,
-) -> AppResult<Json<Value>> {
-    control_routine_session(&state, routine_id, RoutineSessionCommand::Reset).await
-}
-
-pub async fn kill_routine_session(
-    State(state): State<AppState>,
-    Path(routine_id): Path<String>,
-) -> AppResult<Json<Value>> {
-    control_routine_session(&state, routine_id, RoutineSessionCommand::Kill).await
-}
-
-async fn control_routine_session(
-    state: &AppState,
-    routine_id: String,
-    command: RoutineSessionCommand,
-) -> AppResult<Json<Value>> {
-    let store = routine_store(state)?;
-    let Some(routine) = store.get_routine(&routine_id).await.map_err(store_error)? else {
-        return Err(AppError::not_found(format!(
-            "routine {routine_id} not found"
-        )));
-    };
-    if routine.execution_strategy != "persistent" {
-        return Err(AppError::conflict(format!(
-            "routine {routine_id} uses execution_strategy={}; session control requires persistent",
-            routine.execution_strategy
-        )));
-    }
-    if routine
-        .agent_id
-        .as_deref()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .is_none()
-    {
-        return Err(AppError::conflict(format!(
-            "routine {routine_id} is not attached to an agent"
-        )));
-    }
-
-    let reason = format!(
-        "routine persistent session {} via POST /api/routines/{}/session/{}",
-        command.as_str(),
-        routine_id,
-        command.as_str()
-    );
-    let session = routine_session_controller(state)?
-        .control_persistent_session(&routine, command, &reason)
-        .await
-        .map_err(session_control_error)?;
-    let session_changed = session.runtime_cleared
-        || session.tmux_killed
-        || session.inflight_cleared
-        || session.disconnected_sessions > 0;
-    let interrupted_run_id = if session_changed {
-        store
-            .interrupt_in_flight_run(
-                &routine_id,
-                &reason,
-                Some(json!({
-                    "status": "interrupted_by_session_control",
-                    "routine_id": routine_id,
-                    "action": command.as_str(),
-                    "provider": session.provider.clone(),
-                    "channel_id": session.channel_id.clone(),
-                    "session_key": session.session_key.clone(),
-                    "tmux_session": session.tmux_session.clone(),
-                })),
-            )
-            .await
-            .map_err(store_error)?
-    } else {
-        None
-    };
-
-    Ok(Json(json!({
-        "ok": true,
-        "session": session,
-        "interrupted_run_id": interrupted_run_id,
-    })))
-}
-
-fn routine_store(state: &AppState) -> AppResult<RoutineStore> {
-    let Some(pool) = state.pg_pool.clone() else {
-        return Err(AppError::new(
-            StatusCode::SERVICE_UNAVAILABLE,
-            ErrorCode::Database,
-            "postgres pool unavailable; routines require postgresql",
-        ));
-    };
-    Ok(RoutineStore::new_with_timezone_and_checkpoint_limit(
-        std::sync::Arc::new(pool),
-        state.config.routines.default_timezone.clone(),
-        state.config.routines.max_checkpoint_bytes,
-    ))
-}
-
-fn routine_agent_executor(state: &AppState) -> AppResult<RoutineAgentExecutor> {
-    let Some(pool) = state.pg_pool.clone() else {
-        return Err(AppError::new(
-            StatusCode::SERVICE_UNAVAILABLE,
-            ErrorCode::Database,
-            "postgres pool unavailable; routines require postgresql",
-        ));
-    };
-    Ok(RoutineAgentExecutor::new(
-        std::sync::Arc::new(pool),
-        state.health_registry.clone(),
-        state.config.routines.agent_timeout_secs,
-    ))
-}
-
-fn routine_discord_logger(state: &AppState) -> AppResult<RoutineDiscordLogger> {
-    let Some(pool) = state.pg_pool.clone() else {
-        return Err(AppError::new(
-            StatusCode::SERVICE_UNAVAILABLE,
-            ErrorCode::Database,
-            "postgres pool unavailable; routines require postgresql",
-        ));
-    };
-    Ok(RoutineDiscordLogger::new_with_health_registry(
-        std::sync::Arc::new(pool),
-        state.health_registry.clone(),
-        routine_health_target(&state.config),
-    ))
-}
-
-fn routine_session_controller(state: &AppState) -> AppResult<RoutineSessionController> {
-    let Some(pool) = state.pg_pool.clone() else {
-        return Err(AppError::new(
-            StatusCode::SERVICE_UNAVAILABLE,
-            ErrorCode::Database,
-            "postgres pool unavailable; routines require postgresql",
-        ));
-    };
-    Ok(RoutineSessionController::new(
-        std::sync::Arc::new(pool),
-        state.health_registry.clone(),
-    ))
-}
-
-fn fallback_name(script_ref: &str) -> String {
-    std::path::Path::new(script_ref)
-        .file_stem()
-        .unwrap_or_default()
-        .to_string_lossy()
-        .to_string()
-}
-
-fn initial_attach_status(script_ref: &str) -> &'static str {
-    if script_ref == PARALLEL_SAFE_MIGRATED_LAUNCHD_SCRIPT_REF {
-        "enabled"
-    } else if is_migrated_launchd_script_ref(script_ref) {
-        "paused"
-    } else {
-        "enabled"
-    }
-}
-
-fn normalize_script_ref(script_ref: &str) -> AppResult<String> {
-    let normalized = script_ref.trim().replace('\\', "/");
-    if normalized.is_empty() {
-        return Err(AppError::bad_request("script_ref is required"));
-    }
-    Ok(normalized)
-}
-
-fn validate_execution_strategy_request(strategy: &str) -> AppResult<()> {
-    match strategy {
-        "fresh" | "persistent" => Ok(()),
-        other => Err(AppError::bad_request(format!(
-            "unsupported routine execution_strategy '{other}'; expected fresh or persistent"
-        ))),
-    }
-}
-
-fn validate_run_status_filter(status: &str) -> AppResult<()> {
-    match status {
-        "running" | "succeeded" | "failed" | "skipped" | "paused" | "interrupted" => Ok(()),
-        other => Err(AppError::bad_request(format!(
-            "unsupported routine run status '{other}'"
-        ))),
-    }
-}
-
-fn validate_schedule_request(schedule: Option<&str>) -> AppResult<()> {
-    let Some(schedule) = schedule else {
-        return Ok(());
-    };
-    validate_routine_schedule(schedule).map_err(|error| AppError::bad_request(error.to_string()))
-}
-
-fn validate_timeout_request(timeout_secs: Option<i32>) -> AppResult<()> {
-    if matches!(timeout_secs, Some(value) if value <= 0) {
-        return Err(AppError::bad_request(
-            "routine timeout_secs must be greater than zero",
-        ));
-    }
-    Ok(())
-}
-
-fn validate_max_retries_request(max_retries: Option<i32>) -> AppResult<()> {
-    if matches!(max_retries, Some(value) if value < 0) {
-        return Err(AppError::bad_request(
-            "routine max_retries must be greater than or equal to zero",
-        ));
-    }
-    Ok(())
-}
-
-async fn validate_agent_id_request(
-    state: &AppState,
-    field_name: &str,
-    agent_id: Option<&str>,
-) -> AppResult<Option<String>> {
-    let Some(agent_id) = agent_id.map(str::trim).filter(|value| !value.is_empty()) else {
-        return Ok(None);
-    };
-    let Some(pool) = state.pg_pool.as_ref() else {
-        return Err(AppError::new(
-            StatusCode::SERVICE_UNAVAILABLE,
-            ErrorCode::Database,
-            "postgres pool unavailable; routines require postgresql",
-        ));
-    };
-    match crate::db::kanban_cards::resolve_existing_agent_id_with_pg(pool, agent_id).await {
-        Some(existing) => Ok(Some(existing)),
-        None => Err(AppError::bad_request(format!(
-            "routine {field_name} references unknown agent '{agent_id}'"
-        ))),
-    }
-}
-
-fn validate_distinct_fallback_agent(
-    agent_id: Option<&str>,
-    fallback_agent_id: Option<&str>,
-) -> AppResult<()> {
-    let Some(agent_id) = agent_id.map(str::trim).filter(|value| !value.is_empty()) else {
-        return Ok(());
-    };
-    let Some(fallback_agent_id) = fallback_agent_id
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-    else {
-        return Ok(());
-    };
-    if agent_id == fallback_agent_id {
-        return Err(AppError::bad_request(
-            "routine fallback_agent_id must differ from agent_id",
-        ));
-    }
-    Ok(())
-}
-
-fn routine_health_target(config: &Config) -> Option<String> {
-    config
-        .kanban
-        .human_alert_channel_id
-        .as_deref()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(|value| format!("channel:{value}"))
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum RoutineDeleteAuditLevel {
-    Info,
-    Warn,
-}
-
-#[derive(Debug, PartialEq, Eq)]
-struct RoutineDeleteAuditRecord {
-    outcome: &'static str,
-    routine_id: String,
-    owner: String,
-    caller: String,
-    deleted_runs: Option<u64>,
-    level: RoutineDeleteAuditLevel,
-}
-
-fn audit_agent_label(agent_id: Option<&str>, empty_label: &'static str) -> String {
-    agent_id
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .unwrap_or(empty_label)
-        .to_string()
-}
-
-fn routine_delete_audit_record(
-    routine_id: &str,
-    result: &DeleteRoutineResult,
-) -> RoutineDeleteAuditRecord {
-    match result {
-        DeleteRoutineResult::Deleted {
-            run_history_deleted,
-            routine_agent_id,
-            caller_agent_id,
-        } => RoutineDeleteAuditRecord {
-            outcome: "deleted",
-            routine_id: routine_id.to_string(),
-            owner: audit_agent_label(routine_agent_id.as_deref(), "unowned"),
-            caller: audit_agent_label(caller_agent_id.as_deref(), "unresolved"),
-            deleted_runs: Some(*run_history_deleted),
-            level: RoutineDeleteAuditLevel::Info,
-        },
-        DeleteRoutineResult::NotFound { caller_agent_id } => RoutineDeleteAuditRecord {
-            outcome: "not_found",
-            routine_id: routine_id.to_string(),
-            owner: "unknown".to_string(),
-            caller: audit_agent_label(caller_agent_id.as_deref(), "unresolved"),
-            deleted_runs: None,
-            level: RoutineDeleteAuditLevel::Warn,
-        },
-        DeleteRoutineResult::NotDetached {
-            routine_agent_id,
-            caller_agent_id,
-            ..
-        } => RoutineDeleteAuditRecord {
-            outcome: "not_detached",
-            routine_id: routine_id.to_string(),
-            owner: audit_agent_label(routine_agent_id.as_deref(), "unowned"),
-            caller: audit_agent_label(caller_agent_id.as_deref(), "unresolved"),
-            deleted_runs: None,
-            level: RoutineDeleteAuditLevel::Warn,
-        },
-        DeleteRoutineResult::InFlight {
-            routine_agent_id,
-            caller_agent_id,
-        } => RoutineDeleteAuditRecord {
-            outcome: "in_flight",
-            routine_id: routine_id.to_string(),
-            owner: audit_agent_label(routine_agent_id.as_deref(), "unowned"),
-            caller: audit_agent_label(caller_agent_id.as_deref(), "unresolved"),
-            deleted_runs: None,
-            level: RoutineDeleteAuditLevel::Warn,
-        },
-        DeleteRoutineResult::Forbidden {
-            owner,
-            caller_agent_id,
-        } => RoutineDeleteAuditRecord {
-            outcome: "forbidden",
-            routine_id: routine_id.to_string(),
-            owner: audit_agent_label(Some(owner), "unknown"),
-            caller: audit_agent_label(caller_agent_id.as_deref(), "unresolved"),
-            deleted_runs: None,
-            level: RoutineDeleteAuditLevel::Warn,
-        },
-    }
-}
-
-fn audit_routine_delete(routine_id: &str, result: &DeleteRoutineResult) {
-    let record = routine_delete_audit_record(routine_id, result);
-    match (record.level, record.deleted_runs) {
-        (RoutineDeleteAuditLevel::Info, Some(deleted_runs)) => {
-            tracing::info!(
-                event = "routine_delete_audit",
-                outcome = record.outcome,
-                routine_id = %record.routine_id,
-                owner = %record.owner,
-                caller = %record.caller,
-                deleted_runs = deleted_runs,
-                "routine hard delete audited"
-            );
-        }
-        _ => {
-            tracing::warn!(
-                event = "routine_delete_audit",
-                outcome = record.outcome,
-                routine_id = %record.routine_id,
-                owner = %record.owner,
-                caller = %record.caller,
-                "routine hard delete rejected"
-            );
-        }
-    }
-}
-
-fn delete_routine_response(
-    routine_id: &str,
-    result: DeleteRoutineResult,
-) -> AppResult<(StatusCode, Json<Value>)> {
-    match result {
-        DeleteRoutineResult::Deleted {
-            run_history_deleted,
-            ..
-        } => Ok((
-            StatusCode::OK,
-            Json(json!({
-                "ok": true,
-                "routine_id": routine_id,
-                "run_history_deleted": run_history_deleted,
-            })),
-        )),
-        DeleteRoutineResult::NotFound { .. } => Err(AppError::not_found(format!(
-            "routine {routine_id} not found"
-        ))),
-        DeleteRoutineResult::NotDetached { status, .. } => Err(AppError::conflict(format!(
-            "routine {routine_id} must be detached before delete; current status is {status}"
-        ))),
-        DeleteRoutineResult::InFlight { .. } => Err(AppError::conflict(format!(
-            "routine {routine_id} has an in-flight run and cannot be deleted"
-        ))),
-        DeleteRoutineResult::Forbidden {
-            owner,
-            caller_agent_id: Some(caller),
-        } => Err(AppError::new(
-            StatusCode::FORBIDDEN,
-            ErrorCode::Policy,
-            format!(
-                "routine {routine_id} belongs to agent {owner}; caller agent {caller} cannot delete it"
-            ),
-        )),
-        DeleteRoutineResult::Forbidden {
-            owner,
-            caller_agent_id: None,
-        } => Err(AppError::new(
-            StatusCode::FORBIDDEN,
-            ErrorCode::Policy,
-            format!(
-                "routine {routine_id} belongs to agent {owner}; caller agent scope could not be resolved"
-            ),
-        )),
-    }
-}
-
-fn store_error(error: anyhow::Error) -> AppError {
-    if is_resume_routine_requires_next_due_at(&error) {
-        return AppError::conflict(error.to_string());
-    }
-    AppError::internal(error.to_string()).with_code(ErrorCode::Database)
-}
-
-fn session_control_error(error: anyhow::Error) -> AppError {
-    let message = error.to_string();
-    if message.contains("not found") {
-        AppError::not_found(message)
-    } else if message.contains("not configured")
-        || message.contains("not attached")
-        || message.contains("invalid")
-        || message.contains("requires execution_strategy")
-    {
-        AppError::conflict(message)
-    } else {
-        AppError::internal(message)
-    }
-}
+pub use self::handlers::{
+    attach_routine, delete_routine, detach_routine, get_routine, kill_routine_session,
+    list_routine_runs, list_routines, patch_routine, pause_routine, reset_routine_session,
+    resume_routine, routine_metrics, run_routine_now, search_routine_run_results,
+};
 
 #[cfg(test)]
 mod tests {
-    use axum::http::StatusCode;
     use serde_json::json;
 
-    use super::{
-        PARALLEL_SAFE_MIGRATED_LAUNCHD_SCRIPT_REF, PatchRoutineBody, ResumeRoutineBody,
-        RoutineDeleteAuditLevel, delete_routine_response, ensure_routine_runtime_runnable,
-        initial_attach_status, normalize_script_ref, routine_delete_audit_record, store_error,
-        validate_distinct_fallback_agent,
-    };
-    use crate::config::RoutinesConfig;
-    use crate::error::ErrorCode;
-    use crate::services::routines::DeleteRoutineResult;
-
-    #[test]
-    fn run_now_guard_rejects_disabled_routines() {
-        let config = RoutinesConfig::default();
-
-        let err = ensure_routine_runtime_runnable(&config)
-            .expect_err("disabled routines must reject run-now before DB access");
-        assert_eq!(err.status(), StatusCode::SERVICE_UNAVAILABLE);
-        assert_eq!(err.code(), ErrorCode::Config);
-        assert_eq!(err.message(), "routines are disabled by config");
-    }
-
-    #[test]
-    fn fallback_agent_must_differ_from_primary_agent() {
-        assert!(validate_distinct_fallback_agent(Some("codex"), Some("claude")).is_ok());
-        assert!(validate_distinct_fallback_agent(None, Some("claude")).is_ok());
-        assert!(validate_distinct_fallback_agent(Some("codex"), None).is_ok());
-
-        let err = validate_distinct_fallback_agent(Some("codex"), Some("codex"))
-            .expect_err("self fallback must be rejected");
-        assert_eq!(err.status(), StatusCode::BAD_REQUEST);
-    }
-
-    #[test]
-    fn run_now_guard_rejects_invalid_runtime_worker_config() {
-        let mut config = RoutinesConfig {
-            enabled: true,
-            ..RoutinesConfig::default()
-        };
-        config.max_agent_polls_per_tick = 0;
-
-        let err = ensure_routine_runtime_runnable(&config)
-            .expect_err("worker-invalid routines config must reject run-now");
-        assert_eq!(err.status(), StatusCode::SERVICE_UNAVAILABLE);
-        assert_eq!(err.code(), ErrorCode::Config);
-        assert!(
-            err.message().contains("max_agent_polls_per_tick"),
-            "unexpected message: {}",
-            err.message()
-        );
-    }
+    use super::{PatchRoutineBody, ResumeRoutineBody};
 
     #[test]
     fn patch_body_preserves_omitted_nullable_fields() {
@@ -1176,214 +248,5 @@ mod tests {
         let update = body.next_due_at_update().expect("field must be present");
         let ts = update.expect("timestamp must be Some");
         assert_eq!(ts.to_rfc3339(), "2026-04-29T00:00:00+00:00");
-    }
-
-    #[test]
-    fn resume_missing_next_due_store_error_maps_to_conflict() {
-        let err =
-            store_error(crate::services::routines::store::ResumeRoutineRequiresNextDueAt.into());
-        assert_eq!(err.status(), StatusCode::CONFLICT);
-        assert_eq!(
-            err.message(),
-            "next_due_at required to resume schedule-less routine"
-        );
-    }
-
-    #[test]
-    fn delete_routine_response_maps_success_and_conflicts() {
-        let (status, body) = delete_routine_response(
-            "routine-1",
-            DeleteRoutineResult::Deleted {
-                run_history_deleted: 2,
-                routine_agent_id: Some("codex".to_string()),
-                caller_agent_id: Some("codex".to_string()),
-            },
-        )
-        .expect("deleted response");
-        assert_eq!(status, StatusCode::OK);
-        assert_eq!(body.0["ok"], true);
-        assert_eq!(body.0["routine_id"], "routine-1");
-        assert_eq!(body.0["run_history_deleted"], 2);
-
-        let err = delete_routine_response(
-            "routine-1",
-            DeleteRoutineResult::NotDetached {
-                status: "paused".to_string(),
-                routine_agent_id: Some("codex".to_string()),
-                caller_agent_id: Some("codex".to_string()),
-            },
-        )
-        .expect_err("non-detached delete must conflict");
-        assert_eq!(err.status(), StatusCode::CONFLICT);
-        assert_eq!(
-            err.message(),
-            "routine routine-1 must be detached before delete; current status is paused"
-        );
-
-        let err = delete_routine_response(
-            "routine-1",
-            DeleteRoutineResult::InFlight {
-                routine_agent_id: Some("codex".to_string()),
-                caller_agent_id: Some("codex".to_string()),
-            },
-        )
-        .expect_err("in-flight delete must conflict");
-        assert_eq!(err.status(), StatusCode::CONFLICT);
-        assert_eq!(
-            err.message(),
-            "routine routine-1 has an in-flight run and cannot be deleted"
-        );
-    }
-
-    #[test]
-    fn delete_routine_audit_record_covers_all_outcomes() {
-        let record = routine_delete_audit_record(
-            "routine-1",
-            &DeleteRoutineResult::Deleted {
-                run_history_deleted: 2,
-                routine_agent_id: Some("codex".to_string()),
-                caller_agent_id: Some("codex".to_string()),
-            },
-        );
-        assert_eq!(record.outcome, "deleted");
-        assert_eq!(record.routine_id, "routine-1");
-        assert_eq!(record.owner, "codex");
-        assert_eq!(record.caller, "codex");
-        assert_eq!(record.deleted_runs, Some(2));
-        assert_eq!(record.level, RoutineDeleteAuditLevel::Info);
-
-        let record = routine_delete_audit_record(
-            "routine-1",
-            &DeleteRoutineResult::Forbidden {
-                owner: "codex".to_string(),
-                caller_agent_id: Some("claude".to_string()),
-            },
-        );
-        assert_eq!(record.outcome, "forbidden");
-        assert_eq!(record.owner, "codex");
-        assert_eq!(record.caller, "claude");
-        assert_eq!(record.deleted_runs, None);
-        assert_eq!(record.level, RoutineDeleteAuditLevel::Warn);
-
-        let record = routine_delete_audit_record(
-            "routine-1",
-            &DeleteRoutineResult::NotDetached {
-                status: "paused".to_string(),
-                routine_agent_id: Some("codex".to_string()),
-                caller_agent_id: Some("codex".to_string()),
-            },
-        );
-        assert_eq!(record.outcome, "not_detached");
-        assert_eq!(record.owner, "codex");
-        assert_eq!(record.caller, "codex");
-        assert_eq!(record.deleted_runs, None);
-        assert_eq!(record.level, RoutineDeleteAuditLevel::Warn);
-
-        let record = routine_delete_audit_record(
-            "routine-1",
-            &DeleteRoutineResult::InFlight {
-                routine_agent_id: Some("codex".to_string()),
-                caller_agent_id: None,
-            },
-        );
-        assert_eq!(record.outcome, "in_flight");
-        assert_eq!(record.owner, "codex");
-        assert_eq!(record.caller, "unresolved");
-        assert_eq!(record.deleted_runs, None);
-        assert_eq!(record.level, RoutineDeleteAuditLevel::Warn);
-
-        let record = routine_delete_audit_record(
-            "routine-1",
-            &DeleteRoutineResult::NotFound {
-                caller_agent_id: Some("codex".to_string()),
-            },
-        );
-        assert_eq!(record.outcome, "not_found");
-        assert_eq!(record.owner, "unknown");
-        assert_eq!(record.caller, "codex");
-        assert_eq!(record.deleted_runs, None);
-        assert_eq!(record.level, RoutineDeleteAuditLevel::Warn);
-    }
-
-    #[test]
-    fn delete_routine_scope_rejects_other_agent() {
-        let err = delete_routine_response(
-            "routine-1",
-            DeleteRoutineResult::Forbidden {
-                owner: "codex".to_string(),
-                caller_agent_id: Some("claude".to_string()),
-            },
-        )
-        .expect_err("agent-scoped caller must not delete another agent's routine");
-        assert_eq!(err.status(), StatusCode::FORBIDDEN);
-        assert_eq!(err.code(), ErrorCode::Policy);
-        assert_eq!(
-            err.message(),
-            "routine routine-1 belongs to agent codex; caller agent claude cannot delete it"
-        );
-    }
-
-    #[test]
-    fn delete_routine_scope_rejects_unresolved_declared_scope() {
-        let err = delete_routine_response(
-            "routine-1",
-            DeleteRoutineResult::Forbidden {
-                owner: "codex".to_string(),
-                caller_agent_id: None,
-            },
-        )
-        .expect_err("declared but unresolved scope must fail closed");
-        assert_eq!(err.status(), StatusCode::FORBIDDEN);
-        assert_eq!(err.code(), ErrorCode::Policy);
-        assert_eq!(
-            err.message(),
-            "routine routine-1 belongs to agent codex; caller agent scope could not be resolved"
-        );
-    }
-
-    #[test]
-    fn delete_routine_scope_rejects_absent_header_for_owned_routine() {
-        let err = delete_routine_response(
-            "routine-1",
-            DeleteRoutineResult::Forbidden {
-                owner: "codex".to_string(),
-                caller_agent_id: None,
-            },
-        )
-        .expect_err("absent caller scope must fail closed for owned routines");
-        assert_eq!(err.status(), StatusCode::FORBIDDEN);
-        assert_eq!(err.code(), ErrorCode::Policy);
-        assert_eq!(
-            err.message(),
-            "routine routine-1 belongs to agent codex; caller agent scope could not be resolved"
-        );
-    }
-
-    #[test]
-    fn normalize_script_ref_trims_and_matches_loader_separator() {
-        assert_eq!(
-            normalize_script_ref(" nested\\summary.js \n").unwrap(),
-            "nested/summary.js"
-        );
-
-        let err = normalize_script_ref(" \t ").expect_err("empty refs must be rejected");
-        assert_eq!(err.status(), StatusCode::BAD_REQUEST);
-        assert_eq!(err.message(), "script_ref is required");
-    }
-
-    #[test]
-    fn migrated_launchd_attach_defaults_to_paused() {
-        assert_eq!(
-            initial_attach_status("migrated-launchd/memory-merge.js"),
-            "paused"
-        );
-        assert_eq!(
-            initial_attach_status(PARALLEL_SAFE_MIGRATED_LAUNCHD_SCRIPT_REF),
-            "enabled"
-        );
-        assert_eq!(
-            initial_attach_status("monitoring/automation-candidate-detector.js"),
-            "enabled"
-        );
     }
 }
